Level 1 progress (40%)
Added jump animation, fine tuned the animation transition.

Drew and implemented basic scene transition.

Drew animation of instructions but have yet to implement it.

Increased jumpstr

Separated the feet hitbox and the body hitbox to prevent problems such as touching a wall while in mid-air and that being counted as landing on the ground.

In progress of trying to get rid of a hard to achieve bug where while hugging a wall and jumping on top of that wall without losing contact will make the MC be stuck in the jump sprite.